abstracted
