+++ main-site/components/landing-section/LandingSection.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import {
   StyledHackathonText,
   StyledThemeText,
@@ -7,45 +7,20 @@ import {
   StyledToggle
 } from './LandingSection.styles';
 import Toggle from '../../../shared-ui/images/toggle-day-dark.svg';
-import DayBackgroundDesktop from '../../../shared-ui/images/day-background-desktop.svg';
-import NightBackgroundDesktop from '../../../shared-ui/images/dark-background-desktop.svg';
-import DayBackgroundMobile from '../../../shared-ui/images/mobile-light-background.svg';
-import NightBackgroundMobile from '../../../shared-ui/images/mobile-dark-background.svg';
 import Sun from '../../../shared-ui/images/sun.svg';
 import Moon from '../../../shared-ui/images/moon.svg';
-import { min } from '../../../shared-ui/lib/responsive';
-import useMatchMedia from 'react-use-match-media';
-import {
-  StyledBackgrounds,
-  StyledStar
-} from '../../../shared-ui/styled-components/Background.styles';
-import Header from '../../../shared-ui/components/header/Header';
-import { homeTabInfo } from '../../../shared-ui/lib/data';
-import Footer from '../../../shared-ui/components/footer/Footer';
+import { StyledStar } from '../../../shared-ui/styled-components/Background.styles';
+import { LandingSectionProps } from '../../../shared-ui/lib/types';
 
-const LandingSection: React.FC = () => {
-  const [isDay, setIsDay] = useState<boolean>(true);
-  const isDesktop = useMatchMedia(min.tablet);
+const LandingSection: React.FC<LandingSectionProps> = ({ isDay, setIsDay }) => {
   return (
     <>
-      <Header tabs={homeTabInfo} isDay={isDay} />
       <StyledToggle
         src={Toggle}
         alt="toggle-day-night"
         onClick={(): void => setIsDay(!isDay)}
       />
-      {!isDesktop && isDay && (
-        <StyledBackgrounds src={DayBackgroundMobile} />
-      )}
-      {!isDesktop && !isDay && (
-        <StyledBackgrounds src={NightBackgroundMobile} />
-      )}
-       {isDesktop && isDay && (
-        <StyledBackgrounds src={DayBackgroundDesktop} />
-      )}
-      {isDesktop && !isDay && (
-        <StyledBackgrounds src={NightBackgroundDesktop} />
-      )}
+
       {isDay ? <StyledStar src={Sun} /> : <StyledStar src={Moon} />}
       <StyledLandingSectionContainer>
         <StyledHackathonText>HackBeanpot 2023</StyledHackathonText>
@@ -56,10 +31,8 @@ const LandingSection: React.FC = () => {
           newTab
         />
       </StyledLandingSectionContainer>
-      <Footer tabs={homeTabInfo} isDay={isDay} />
     </>
   );
 };
 
 export default LandingSection;
-
